requirements.txt
asgiref==3.7.2
comtypes==1.2.0
Django==4.2.6
django-cors-headers==4.3.1
django-filter==23.4
django-jazzmin==2.6.0
djangorestframework==3.14.0
Pillow==10.1.0
pypiwin32==223
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
